今天完成了文章后台查询的查询功能和删除功能 修改了一堆bug，规范了数据库与实体类的设计 完善了数据长度对应类型的修改和日志

+++ src/main/java/com/codermy/myspringsecurityplus/fore/dao/ArticleDao.java
@@ -5,6 +5,7 @@ import org.apache.ibatis.annotations.Insert;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Select;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Mapper
@@ -39,7 +40,7 @@ public interface ArticleDao {
      * @param ids 需要删除的数据ID
      * @return 结果
      */
-    int deleteArticleByIds(Integer[] ids);
+    int deleteArticleByIds(Long[] ids);
 
 //    /**
 //     * 根据用户ID查询文章
